Add files via upload
This is our protagonist of the comics, named Agnes. She is not a specific user of CleverCogs, but a persona who best represents the most typical characteristics of all users in terms of age, gender, locations, and medical conditions. So, in this part, we also present the users’ demographic data as supporting evidence for our character settings.