Add files via upload
This graph shows Agnes's activities before and now at 3pm. After moving to CleverCogs, the data shows that more people are using Youtube between 3pm and 5pm. If you click on Gender, Age, Location, and Health Condition below, you can also see the data characteristics of the people who are using this link at the same time. So, this is a recommendation system based on data analysis.