Add files via upload
In this picture, I analyzed the total platform accessed of CleverCogs over the past four years (2018-2021). And in the next work, I also use gender to distinguish them, and respectively conduct accessed data display and comparison. However, it is worth noting that the number of link visits does not represent a user's individual choice. Therefore, when analyzing the actual usage of the user, it is necessary to remove the duplicate ID value to draw a more accurate conclusion of the network activity preference.